// src/utils/enums.ts
export enum COLOR_CODE {
  BLACK = '#000000',
  WHITE = '#FFFFFF',
  RED = '#C00000',
  ACCENT = '#2574FB',
  GALLERY_WHITE = '#ECECEC',
  SILVER_CHALICE = '#9E9E9E',
  WHITE_LILAC = '#F6F7FC',
  MINE_SHAFT = '#222222',
  DUSTY_GRAY = '#9A9A9A',
  JADE = '#00BC78',
  MANDY = '#EE4B69',
}

export enum FILE_UPLOAD_STATUS {
  UPLOADING = 'UPLOADING',
  NEXT_UP = 'NEXT_UP',
  COMPLETED = 'COMPLETED',
  INCOMPLETE = 'INCOMPLETE'
}